requirements.txt
python-binance==1.0.19
pandas==2.1.4
python-telegram-bot==20.7
python-dotenv==1.0.0

numpy==1.26.3
matplotlib==3.8.2
plotly==5.18.0 